pyproject.toml
[project]
name = "fast-project"
version = "0.1.0"
description = "um breve e pequeno projeto com o fastapi"
authors = [
    { name = "Your Name", email = "vald.soares360@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

dependencies = [
    "fastapi[standard]>=0.128.0,<0.129.0",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "tzdata (>=2025.3,<2026.0)"
]

[tool.poetry]
packages = [{ include = "fast_project", from = "src" }]

[dependency-groups]
dev = [
    "ruff>=0.14.14,<0.15.0",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]

[tool.pytest.ini_options]
pythonpath = "."  
addopts = '-p no:warnings'

[tool.ruff.format]
preview = true
quote-style = "single"


[tool.taskipy.tasks]
lint = 'ruff check'
format = 'ruff format   '
run = 'fastapi dev src/fast_project/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src/fast_project -vv'   
post_test = 'coverage html'
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
